fix: open stickers settings section from nav and composer
isSettingsSection() validates /settings?section= against
SETTINGS_SECTION_VALUES, which was missing "stickers" — the nav
descriptor and render case existed, but the route search validation
stripped the param and fell back to the default profile section, so
clicking Stickers in settings (or "Open sticker settings" in the
composer picker) never opened the panel.

Extract the section registry into features/settings/lib/settingsSections.ts
(with "stickers" added) and re-export it from SettingsPanels.tsx, plus a
regression test asserting every nav section passes the guard.
