+++ desktop/src/features/settings/lib/settingsSections.ts
@@ -1,0 +1,45 @@
+export type SettingsSection =
+  | "profile"
+  | "notifications"
+  | "experimental"
+  | "agents"
+  | "channel-templates"
+  | "compute"
+  | "appearance"
+  | "shortcuts"
+  | "hosted-communities"
+  | "community-members"
+  | "moderation"
+  | "custom-emoji"
+  | "stickers"
+  | "local-archive"
+  | "mobile"
+  | "updates";
+
+export const DEFAULT_SETTINGS_SECTION: SettingsSection = "profile";
+
+const SETTINGS_SECTION_VALUES: readonly SettingsSection[] = [
+  "profile",
+  "notifications",
+  "experimental",
+  "agents",
+  "channel-templates",
+  "compute",
+  "appearance",
+  "shortcuts",
+  "hosted-communities",
+  "community-members",
+  "moderation",
+  "custom-emoji",
+  "stickers",
+  "local-archive",
+  "mobile",
+  "updates",
+];
+
+export function isSettingsSection(value: unknown): value is SettingsSection {
+  return (
+    typeof value === "string" &&
+    (SETTINGS_SECTION_VALUES as readonly string[]).includes(value)
+  );
+}
