+++ desktop/src/features/settings/ui/SettingsPanels.tsx
@@ -87,50 +87,12 @@ import { ProfileSettingsCard } from "./ProfileSettingsCard";
 import { UpdateChecker } from "../UpdateChecker";
 import { SettingsSectionHeader } from "./SettingsSectionHeader";
 
-export type SettingsSection =
-  | "profile"
-  | "notifications"
-  | "experimental"
-  | "agents"
-  | "channel-templates"
-  | "compute"
-  | "appearance"
-  | "shortcuts"
-  | "hosted-communities"
-  | "community-members"
-  | "moderation"
-  | "custom-emoji"
-  | "stickers"
-  | "local-archive"
-  | "mobile"
-  | "updates";
-
-export const DEFAULT_SETTINGS_SECTION: SettingsSection = "profile";
-
-const SETTINGS_SECTION_VALUES: readonly SettingsSection[] = [
-  "profile",
-  "notifications",
-  "experimental",
-  "agents",
-  "channel-templates",
-  "compute",
-  "appearance",
-  "shortcuts",
-  "hosted-communities",
-  "community-members",
-  "moderation",
-  "custom-emoji",
-  "local-archive",
-  "mobile",
-  "updates",
-];
-
-export function isSettingsSection(value: unknown): value is SettingsSection {
-  return (
-    typeof value === "string" &&
-    (SETTINGS_SECTION_VALUES as readonly string[]).includes(value)
-  );
-}
+import type { SettingsSection } from "../lib/settingsSections";
+export {
+  DEFAULT_SETTINGS_SECTION,
+  isSettingsSection,
+  type SettingsSection,
+} from "../lib/settingsSections";
 
 export type SettingsSectionDescriptor = {
   value: SettingsSection;
